mfbm/requirements.txt
bleach==1.5.0
cycler==0.10.0
Cython==0.27.3
decorator==4.1.2
enum34==1.1.6
fbm==0.1.1
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.2
Markdown==2.6.9
matplotlib==2.1.0
natsort==5.1.1
networkx==2.0
numpy==1.13.3+mkl
olefile==0.44
pandas==0.21.0
Pillow==4.3.0
pip==9.0.1
protobuf==3.5.0.post1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0rc1
setuptools==38.2.3
six==1.11.0
sklearn==0.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
Werkzeug==0.12.2
wheel==0.30.0
